Updated the search bar
The search bar can link to the search screen by press the search icon on the right. But I don't know how to use the api to this, so I will leave it like this. The search bar name "SearchTop" in the components folder, and the search screen name "SearchList" in the screens folder.